+++ lib/modules/network/network_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: network
-// Updated at: 2026-05-29 13:07:07
-// Commit: test(network): implement widget integration tests for network
+// Updated at: 2026-05-29 13:51:55
+// Commit: fix(network): prevent double tap duplicate action in network
 
 class networkComponentState {
-  final String commitRef = '1576';
-  final String lastMessage = 'test(network): implement widget integration tests for network';
-  final DateTime timestamp = DateTime.parse('2026-05-29T13:07:07+05:00');
+  final String commitRef = '1582';
+  final String lastMessage = 'fix(network): prevent double tap duplicate action in network';
+  final DateTime timestamp = DateTime.parse('2026-05-29T13:51:55+05:00');
 }